+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.1] - 2023-02-06
+
+### Fixed
+- Keep `rawXml` in `parseXmlResponse` result
+- Generic return type of `parseXmlResponse`
+
 ## [3.3.0] – 2023-01-24
 
 ### Added